move TensorHelper to tensor_indices.h
